# openapi.yaml
openapi: '3.0.3'
info:
  title: Multi-Tenant Project Management API
  description: |
    RESTful API for a multi-tenant project management system with tenant isolation via API keys,
    tamper-evident audit trails, sliding window rate limiting, and queue-based email engine.
  version: '1.0.0'
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: 'API key in format: vz_xxxx'
    InternalKey:
      type: apiKey
      in: header
      name: x-internal-key
...
paths:
...
  /workspaces:
    get:
      summary: List workspaces for current tenant
      tags: [Workspaces]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of workspaces
    post:
      summary: Create workspace
      tags: [Workspaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Workspace created
  /workspaces/{workspaceId}:
    get:
      summary: Get workspace details
      tags: [Workspaces]
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace details
    put:
      summary: Update workspace (owner only)
      tags: [Workspaces]
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace updated
  /workspaces/{workspaceId}/members:
    get:
      summary: List workspace members
      tags: [Workspaces]
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of members
    post:
      summary: Add member to workspace (sends invite email)
      tags: [Workspaces]
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId: { type: string }
                role: { type: string, enum: [owner, member] }
      responses:
        '201':
          description: Member added and invite email sent
...